12-005 Preventing Unnecessary Re-Evaluations with React.memo()
React.memo( )  can  check the props and see if it has been updated.
if not , then the component will skip the re-evaluation.

but for those reference type, Such as Functions, objects.  if the parent component changes, these functions or objects will be recreated a brand-new one.  Even thought they are doing the same thing as the previous function or object, but they are as same as the old one.   so that reference type will be regarded as updated..... React.memo()  cannot deal with this situation.