12-011 Optimizing with useMemo
here, the sort function will depends on "items" ,   which is an array, and a reference type.
even if an Array keep the same items,   as it is reference type, an new  created array is always different.

so that to make the Array looks like no-change, we can still use the useMemo( )  to reserve it .